+++ src/app/conge/conge.component.ts
@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { Conge } from 'src/Modeles/Conge';
 import { Enseignant } from 'src/Modeles/Enseignant';
 import { CongeService } from 'src/Services/conge.service';
 import { EnseignantService } from 'src/Services/enseignant.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-conge',
@@ -18,7 +20,7 @@ export class CongeComponent implements OnInit {
   constructor(
     private congeService: CongeService,
     private enseignantService: EnseignantService,
-    private router: Router
+    private router: Router,private dialog: MatDialog
   ) { }
 
   ngOnInit() {
@@ -52,15 +54,24 @@ export class CongeComponent implements OnInit {
     this.router.navigate(['/edit-conge', id]);
   }
 
-  deleteConge(id: number) {
-    this.congeService.deleteConge(id).subscribe(
-      () => {
-        this.getConges();
-      },
-      (error) => {
-        console.error('Erreur lors de la suppression du congé :', error);
+  deleteConge(id: number) :void{
+    let dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      height: '200px',
+      width: '300px',
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.congeService.deleteConge(id).subscribe(
+          () => {
+            this.getConges();
+          },
+          (error) => {
+            console.error('Erreur lors de la suppression du congé :', error);
+          }
+        );
       }
-    );
+    });
+   
   }
 
   getEnseignantName(idEnseignant: number): string {
