+++ src/app/niveau/niveau.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { Niveau } from 'src/Modeles/Niveau';
 import { NiveauService } from 'src/Services/niveau.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-niveau',
@@ -12,7 +14,7 @@ export class NiveauComponent implements OnInit {
   niveaux: Niveau[] = [];
   displayedColumns: string[] = ['1', '2', '3'];
 
-  constructor(private niveauService: NiveauService, private router: Router) { }
+  constructor(private niveauService: NiveauService, private router: Router,private dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.getAllNiveaux();
@@ -23,9 +25,18 @@ export class NiveauComponent implements OnInit {
   }
 
   deleteNiveau(id: number): void {
-    this.niveauService.deleteNiveau(id).subscribe(() => {
-      this.niveaux = this.niveaux.filter(niveau => niveau.id !== id);
+    let dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      height: '200px',
+      width: '300px',
     });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.niveauService.deleteNiveau(id).subscribe(() => {
+          this.niveaux = this.niveaux.filter(niveau => niveau.id !== id);
+        });
+      }
+    });
+    
   }
 
   editNiveau(niveau: Niveau): void {
