+++ src/app/salle/salle.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Salle } from 'src/Modeles/Salle';
 import { SalleService } from 'src/Services/salle.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-salle',
@@ -12,7 +14,7 @@ export class SalleComponent  implements OnInit{
   salles: Salle[] = [];
   displayedColumns: string[] = ['1', '2', '3', '4'];
 
-  constructor(private salleService: SalleService, private router: Router) { }
+  constructor(private salleService: SalleService, private router: Router, private dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.getAllSalles();
@@ -23,9 +25,18 @@ export class SalleComponent  implements OnInit{
   }
 
   deleteSalle(id: number): void {
-    this.salleService.deleteSalle(id).subscribe(() => {
-      this.salles = this.salles.filter(salle => salle.id !== id);
+    let dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      height: '200px',
+      width: '300px',
     });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.salleService.deleteSalle(id).subscribe(() => {
+          this.salles = this.salles.filter(salle => salle.id !== id);
+        });
+      }
+    });
+    
   }
 
   editSalle(id: number): void {
